Copy of note to the group.  I have not yet updated the manual or test suite.
I made some changes to interned tries that I think will make them
safer and easier to use, and maybe even speed them up a little.
However, there are some small changes in behavior.

Recall that, for now, interned tries must be private in the MT engine
(I'm looking to change this, but that's how they are now).

The changes are steps towards an interned trie garbage collector, but
todays update is only part of the way there.  Still, I think it is a
good step forward.

The first change occurs when a term is interned into a trie $T$.
Recall that when this happens we may make checks to determine whether
to hash the children of a particular node in the trie, or to resize
the hash table of that node.  If either of these checks are positive,
I made a change to run the choice point stack to see whether there are
any choice points for nodes in $T$, in which case it is not safe to
rehash.  If there are trie choice points for tables, asserted tries,
or interned tries other than $T$it doesn't matter -- its safe to
rehash.  Under this strategy we'll will recheck whenever the next
insert hits that node.  The idea is that adding an extra node wont
slow down access time considerably, and we'll always have the chance
to recheck if we expand the sibling chain (or fill up the hash table)
even further.

Similarly, I made a change in trie_unintern/2 (which reclaims space).
This runs the choice point stack with the same check as above, and if
its safe to reclaim space we do, otherwise we mark the node as deleted
just as in trie_unintern_nr/2.

Finally, I made a change to delete_trie/1 to check the choice point
stack and throw an error if there are choice points into the trie to
be deleted.

The positive side of this is fewer core dumps and the ability to use
interned tries in more types of applications.  The down side of this
is that we may be running the choice point stack more than we'd like.

To address this, first note that if the choice point stack is small,
this isn't a problem -- see the timings below.  However, to avoid
multiple checks during deletion, I added two new predicates,
trie_retractall/2, which does one check of the choice point stack,
then does multiple non-checked "retracts", with or without space
reclamation depending on the check.  Similarly, I added a new
predicate trie_bulk_insert/2 which allows a single check followed by
multiple inserts (with or without a check).

On my Mac, all tests are on an interned trie of 8 million facts.

------------------------------------------------------------
Interning facts in tries

Before adding the CPS checks:
To intern 8 million facts as below took 6.6 secs with trie_intern/4 (via time_create/2)
To intern 8 million facts as below took 4.2 secs with trie_intern/2 (via time_create/2)

After adding the CPS checks:
To intern 8 million facts as below took 4.8 secs with trie_intern/2 (via time_create/2)
To intern 8 million facts as below took 4.3 secs with trie_intern/2 (via time_create/2)

So when the CPS stack is small, you dont pay a price for hashing, but
you might if the CPS stack were big enough.

------------------------------------------------------------
Traversing tries

To traverse the trie, without doing anything other than unification
2.7 secs (using trie_interned/4 in time_traverse/1)

To traverse the trie, without doing anything other than unification
0.7 secs (using trie_interned/2 in time_traverse/1)

------------------------------------------------------------
To delete the trie with delete_trie takes 0.6 seconds (via
time_delete)

------------------------------------------------------------
To retractall the trie via trie_retractall/1 took 7.9 seconds (i.e. by
retracting p(_,_,_)

------------------------------------------------------------
To retractall 1/200th of the trie via trie_retractall/1 took 0.4 seconds (i.e. by
retracting p(1,_,_)

+++ emu/tr_utils.c
@@ -63,6 +63,7 @@
 #include "storage_xsb.h"
 #include "hash_xsb.h"
 #include "tables.h"
+#include "builtin.h"
 
 #include "call_graph_xsb.h" /* incremental evaluation */
 
@@ -1325,24 +1326,26 @@ Integer newtrie(CTXTdecl)
 }
 
 /*----------------------------------------------------------------------*/
+/* i_trie_intern(_Term,_Root,_Leaf,_Flag,_Check_CPS,_Expand_or_not)     */
 
 void trie_intern(CTXTdecl)
 {
   prolog_term term;
   int RootIndex;
-  int flag;
+  int flag, check_cps_flag, expand_flag;
   BTNptr Leaf;
 
   term = ptoc_tag(CTXTc 1);
   RootIndex = ptoc_int(CTXTc 2);
-
+  check_cps_flag = ptoc_int(CTXTc 5);
+  expand_flag = ptoc_int(CTXTc 6);
   xsb_dbgmsg((LOG_INTERN, "Interning "));
   dbg_printterm(LOG_INTERN,stddbg,term,25);
   xsb_dbgmsg((LOG_INTERN, "In trie with root %d", RootIndex));
 
   switch_to_trie_assert;
   SYS_MUTEX_LOCK(MUTEX_TRIE);
-  Leaf = whole_term_chk_ins(CTXTc term,&(Set_ArrayPtr[RootIndex]),&flag);
+  Leaf = whole_term_chk_ins(CTXTc term,&(Set_ArrayPtr[RootIndex]),&flag,check_cps_flag,expand_flag);
   SYS_MUTEX_UNLOCK(MUTEX_TRIE);
   switch_from_trie_assert;
   
@@ -1417,15 +1420,69 @@ void trie_dispose(CTXTdecl)
 {
   BTNptr Leaf;
   long Rootidx;
+  int disposalType;
 
   Rootidx = ptoc_int(CTXTc 1);
   Leaf = (BTNptr)ptoc_int(CTXTc 2);
+  disposalType = ptoc_int(CTXTc 3);
   SYS_MUTEX_LOCK(MUTEX_TRIE);
   switch_to_trie_assert;
   SYS_MUTEX_UNLOCK(MUTEX_TRIE);
-  delete_branch(CTXTc Leaf, &(Set_ArrayPtr[Rootidx]));
+  if (disposalType == NO_CPS_CHECK)     delete_branch(CTXTc Leaf, &(Set_ArrayPtr[Rootidx]));
+  else {
+    if (!interned_trie_cps_check(CTXTc Set_ArrayPtr[Rootidx])) {
+      //          printf(" really deleting branch \n");
+      delete_branch(CTXTc Leaf, &(Set_ArrayPtr[Rootidx]));
+    }
+    else {
+      //          printf(" safely deleting branch \n");
+      safe_delete_branch(Set_ArrayPtr[Rootidx]);
+    }
+  }
   switch_from_trie_assert;
 }
+
+/*----------------------------------------------------------------------*/
+
+#define is_trie_instruction(cp_inst) \
+ ((int) cp_inst >= 0x5c && (int) cp_inst < 0x80) \
+	   || ((int) cp_inst >= 0x90 && (int) cp_inst < 0x94) 
+
+int interned_trie_cps_check(CTXTdeclc BTNptr root) 
+{
+  CPtr cp_top1,cp_bot1 ;
+  byte cp_inst;
+  int found_interned;
+  BTNptr pLeaf;
+
+  cp_bot1 = (CPtr)(tcpstack.high) - CP_SIZE;
+
+  cp_top1 = breg ;				 
+  found_interned = 0;
+  while ( cp_top1 < cp_bot1 && !(found_interned)) {
+    cp_inst = *(byte *)*cp_top1;
+    // Want trie insts, but will need to distinguish from
+    // asserted and interned tries
+    if ( is_trie_instruction(cp_inst) ) {
+      // Below we want basic_answer_trie_tt, ts_answer_trie_tt
+      if (IsInInternTrie(((BTNptr) *cp_top1))) {
+	//	printf(" found interned trie instruction\n");
+	pLeaf = (BTNptr) *cp_top1;
+	while ( IsNonNULL(pLeaf) && (! IsTrieRoot(pLeaf)) && 
+		((int) TN_Instr(pLeaf) != trie_fail_unlock) ) {
+	  pLeaf = BTN_Parent(pLeaf);
+	}
+	if (pLeaf == root) {
+	  //	  printf(" found root!\n");
+	  found_interned = 1;
+	}
+      }
+    }
+    cp_top1 = cp_prevtop(cp_top1);
+  }
+  return found_interned;
+}
+
 
 /*----------------------------------------------------------------------*/
 
@@ -1714,10 +1771,6 @@ void trie_undispose(CTXTdeclc long rootIdx, BTNptr leafn)
 
 /* used by mt engine for shared tables */
 DelTFptr deltf_chain_begin = (DelTFptr) NULL;
-
-#define is_trie_instruction(cp_inst) \
- ((int) cp_inst >= 0x5c && (int) cp_inst < 0x80) \
-	   || ((int) cp_inst >= 0x90 && (int) cp_inst < 0x94) 
 
 /* - - - - - */
